@@ -4,7 +4,7 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": ">8.0.2",
+        "php": ">8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "composer/package-versions-deprecated": "1.11.99.2",
@@ -13,23 +13,22 @@
         "doctrine/orm": "^2.9",
         "erusev/parsedown": "^1.7",
         "sensio/framework-extra-bundle": "^6.1",
-        "symfony/console": "5.3.*",
-        "symfony/dependency-injection": "5.3.*",
-        "symfony/dotenv": "5.3.*",
+        "symfony/console": "6.0.*",
+        "symfony/dependency-injection": "6.0.*",
+        "symfony/dotenv": "6.0.*",
         "symfony/flex": "^1.3.1",
-        "symfony/form": "5.3.*",
-        "symfony/framework-bundle": "5.3.*",
-        "symfony/mailer": "5.3.*",
-        "symfony/proxy-manager-bridge": "5.3.*",
+        "symfony/form": "6.0.*",
+        "symfony/framework-bundle": "6.0.*",
+        "symfony/mailer": "6.0.*",
+        "symfony/proxy-manager-bridge": "6.0.*",
         "symfony/requirements-checker": "^2.0",
-        "symfony/runtime": "5.3.*",
-        "symfony/security-bundle": "5.3.*",
-        "symfony/twig-bundle": "5.3.*",
-        "symfony/ux-dropzone": "^1.3",
-        "symfony/validator": "5.3.*",
-        "symfony/var-dumper": "5.3.*",
+        "symfony/runtime": "6.0.*",
+        "symfony/security-bundle": "6.0.*",
+        "symfony/twig-bundle": "6.0.*",
+        "symfony/validator": "6.0.*",
+        "symfony/var-dumper": "6.0.*",
         "symfony/webpack-encore-bundle": "^1.12",
-        "symfony/yaml": "5.3.*",
+        "symfony/yaml": "6.0.*",
         "twig/extra-bundle": "^3.3",
         "twig/markdown-extra": "^3.3",
         "twig/string-extra": "^3.3",
@@ -81,13 +80,13 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.3.*"
+            "require": "6.0.*"
         }
     },
     "require-dev": {
         "doctrine/doctrine-fixtures-bundle": "^3.4",
         "symfony/maker-bundle": "^1.33",
-        "symfony/stopwatch": "^5.3",
-        "symfony/web-profiler-bundle": "^5.3"
+        "symfony/stopwatch": "^6.0",
+        "symfony/web-profiler-bundle": "^6.0"
     }
 }
